requirements.txt
streamlit==1.32.0
azure-cognitiveservices-vision-computervision==0.9.0
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
pytesseract==0.3.10
numpy==1.26.4 